part 4 finished

+++ part4/bloglist-fullstack/bloglist-backend/controllers/blogs.js
@@ -1,26 +1,26 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
+
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
+    const user = request.user
 
+    if (!user) {
+        response.status(401).send("Unauthorized").end()
+        return
+    }
 
     if (!body.title || !body.author) {
         response.status(400).send("missing title or author").end()
         return
     }
-
-
-    // It iis wad id iis for now
-    const userId = "64a51e7ce3df191c6028efb3"
-
-    const user = await User.findById(userId)
 
     const blog = new Blog({
         title: body.title,
@@ -38,17 +38,32 @@ blogsRouter.post('/', async (request, response) => {
     response.status(201).json(savedBlog)
 })
 
+
 blogsRouter.delete('/:id', async (request, response) => {
     if (!request.params.id) {
         response.status(400).end()
         return
     }
 
-    const deletedBlog = await Blog.findByIdAndRemove(request.params.id)
-    response.json(deletedBlog)
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(404).send("blog not found").end()
+    }
+
+    const user = request.user
+
+    if (user.id === blog.user.toString()) {
+        const deletedBlog = await Blog.findByIdAndRemove(request.params.id)
+        response.json(deletedBlog)
+    } else{
+        response.status(400).end()
+        return
+    }
+
 })
 
-// patch
+
 blogsRouter.put('/:id', async (request, response) => {
     const body = request.body
